requirements.txt
apispec==6.8.1
apispec-webframeworks==1.2.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Deprecated==1.2.18
ecdsa==0.19.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
flask-redoc==0.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==4.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
mdurl==0.1.2
mysql-connector-python==9.2.0
ordered-set==4.1.0
packaging==24.2
pyasn1==0.4.8
Pygments==2.19.1
python-jose==3.4.0
PyYAML==6.0.2
referencing==0.36.2
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
